move core.config to _config.config, update imports
