+++ pandas/_libs/tslibs/parsing.pyx
@@ -30,6 +30,8 @@ from dateutil.relativedelta import relativedelta
 from dateutil.parser import DEFAULTPARSER
 from dateutil.parser import parse as du_parse
 
+from pandas._config.config import get_option
+
 from pandas._libs.tslibs.ccalendar import MONTH_NUMBERS
 from pandas._libs.tslibs.nattype import nat_strings, NaT
 
@@ -50,20 +52,6 @@ cdef:
     set _not_datelike_strings = {'a', 'A', 'm', 'M', 'p', 'P', 't', 'T'}
 
 # ----------------------------------------------------------------------
-
-_get_option = None
-
-
-def get_option(param):
-    """ Defer import of get_option to break an import cycle that caused
-    significant performance degradation in Period construction. See
-    GH#24118 for details
-    """
-    global _get_option
-    if _get_option is None:
-        from pandas.core.config import get_option
-        _get_option = get_option
-    return _get_option(param)
 
 
 def parse_datetime_string(date_string, freq=None, dayfirst=False,
