+++ pandas/core/api.py
@@ -48,7 +48,7 @@ from pandas.tseries.offsets import DateOffset
 from pandas.core.tools.datetimes import to_datetime
 from pandas.core.tools.timedeltas import to_timedelta
 
-from pandas.core.config import (get_option, set_option, reset_option,
+from pandas._config.config import (get_option, set_option, reset_option,
                                 describe_option, option_context, options)
