+++ pandas/core/computation/expressions.py
@@ -14,7 +14,7 @@ from pandas.core.dtypes.generic import ABCDataFrame
 
 import pandas.core.common as com
 from pandas.core.computation.check import _NUMEXPR_INSTALLED
-from pandas.core.config import get_option
+from pandas._config.config import get_option
 
 if _NUMEXPR_INSTALLED:
     import numexpr as ne
