+++ pandas/core/config_init.py
@@ -9,8 +9,8 @@ If you need to make sure options are available even before a certain
 module is imported, register them here rather then in the module.
 
 """
-import pandas.core.config as cf
-from pandas.core.config import (
+import pandas._config.config as cf
+from pandas._config.config import (
     is_bool, is_callable, is_instance_factory, is_int, is_one_of_factory,
     is_text)
