+++ pandas/core/dtypes/dtypes.py
@@ -138,7 +138,7 @@ class PandasExtensionDtype(_DtypeOpsMixin):
         Invoked by bytes(obj) in py3 only.
         Yields a bytestring in both py2/py3.
         """
-        from pandas.core.config import get_option
+        from pandas._config.config import get_option
 
         encoding = get_option("display.encoding")
         return self.__unicode__().encode(encoding, 'replace')
