+++ pandas/core/dtypes/missing.py
@@ -172,7 +172,7 @@ def _use_inf_as_na(key):
     * http://stackoverflow.com/questions/4859217/
       programmatically-creating-variables-in-python/4859312#4859312
     """
-    from pandas.core.config import get_option
+    from pandas._config.config import get_option
     flag = get_option(key)
     if flag:
         globals()['_isna'] = _isna_old
