+++ pandas/io/excel/_base.py
@@ -13,7 +13,7 @@ from pandas.util._decorators import Appender, deprecate_kwarg
 from pandas.core.dtypes.common import (
     is_bool, is_float, is_integer, is_list_like)
 
-from pandas.core import config
+from pandas._config import config
 from pandas.core.frame import DataFrame
 
 from pandas.io.common import _NA_VALUES, _stringify_path, _validate_header_arg
