+++ pandas/io/formats/format.py
@@ -27,7 +27,7 @@ from pandas.core.dtypes.missing import isna, notna
 from pandas import compat
 from pandas.core.base import PandasObject
 import pandas.core.common as com
-from pandas.core.config import get_option, set_option
+from pandas._config.config import get_option, set_option
 from pandas.core.index import Index, ensure_index
 from pandas.core.indexes.datetimes import DatetimeIndex
