+++ pandas/io/formats/html.py
@@ -13,7 +13,7 @@ from pandas.compat import lzip, map, range, u, unichr, zip
 from pandas.core.dtypes.generic import ABCMultiIndex
 
 from pandas import compat, option_context
-from pandas.core.config import get_option
+from pandas._config.config import get_option
 
 from pandas.io.common import _is_url
 from pandas.io.formats.format import TableFormatter, get_level_lengths
