+++ pandas/io/formats/style.py
@@ -21,7 +21,7 @@ from pandas.core.dtypes.generic import ABCSeries
 import pandas as pd
 from pandas.api.types import is_dict_like, is_list_like
 import pandas.core.common as com
-from pandas.core.config import get_option
+from pandas._config.config import get_option
 from pandas.core.generic import _shared_docs
 from pandas.core.indexing import _maybe_numeric_slice, _non_reducing_slice
