+++ pandas/tests/indexes/multi/test_format.py
@@ -54,7 +54,7 @@ def test_format_sparse_display():
 
 
 def test_repr_with_unicode_data():
-    with pd.core.config.option_context("display.encoding", 'UTF-8'):
+    with pd._config.config.option_context("display.encoding", 'UTF-8'):
         d = {"a": [u("\u05d0"), 2, 3], "b": [4, 5, 6], "c": [7, 8, 9]}
         index = pd.DataFrame(d).set_index(["a", "b"]).index
         assert "\\u" not in repr(index)  # we don't want unicode-escaped
