+++ pandas/tests/indexes/test_category.py
@@ -3,6 +3,8 @@
 import numpy as np
 import pytest
 
+import pandas._config.config as cf
+
 from pandas._libs import index as libindex
 from pandas.compat import PY3, range
 
@@ -10,7 +12,6 @@ from pandas.core.dtypes.dtypes import CategoricalDtype
 
 import pandas as pd
 from pandas import Categorical, IntervalIndex, compat
-import pandas.core.config as cf
 from pandas.core.indexes.api import CategoricalIndex, Index
 import pandas.util.testing as tm
 from pandas.util.testing import assert_almost_equal
