+++ pandas/tests/plotting/test_converter.py
@@ -5,11 +5,12 @@ import sys
 import numpy as np
 import pytest
 
+import pandas._config.config as cf
+
 from pandas.compat import u
 from pandas.compat.numpy import np_datetime64_compat
 
 from pandas import Index, Period, Series, Timestamp, date_range
-import pandas.core.config as cf
 import pandas.util.testing as tm
 
 from pandas.tseries.offsets import Day, Micro, Milli, Second
